Move SiteMessageBar to example-app-base
