+++ packages/example-app-base/src/components/SiteMessageBar/index.ts
@@ -1,0 +1,2 @@
+export * from './SiteMessageBar';
+export * from './SiteMessageBar.types';
